feat: phase 1+2 finance and dashboards
- Red-neutral theme, form action layout polish
- Remove admin bank verification flow (soft transition)
- Standardize money to VND integer end-to-end (DB/entity/DTO/UI)
- Add method-level RBAC with @PreAuthorize and update tests
- Phase 2: per-student session tuition line-items + tutor roster view
- Admin: tutor management actions (revoke tutor role, class count, payout actions)

Made-with: Cursor

+++ src/main/java/com/example/tms/api/dto/classes/AvailableClassResponse.java
@@ -1,6 +1,5 @@
 package com.example.tms.api.dto.classes;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -8,7 +7,7 @@ public record AvailableClassResponse(
         UUID classId,
         String displayName,
         String subjectName,
-        BigDecimal pricePerHour,
+        Long pricePerHour,
         String note,
         List<String> studentNames,
         boolean hasApplied
