+++ src/main/java/com/example/tms/api/dto/classes/SubjectOptionResponse.java
@@ -1,11 +1,10 @@
 package com.example.tms.api.dto.classes;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record SubjectOptionResponse(
         UUID id,
         String name,
-        BigDecimal defaultPricePerHour
+        Long defaultPricePerHour
 ) {
 }
