+++ src/main/java/com/example/tms/api/dto/dashboard/TutorDashboardResponse.java
@@ -1,12 +1,10 @@
 package com.example.tms.api.dto.dashboard;
 
-import java.math.BigDecimal;
-
 public record TutorDashboardResponse(
         int year,
         int month,
-        BigDecimal grossRevenue,
-        BigDecimal netSalary,
+        Long grossRevenue,
+        Long netSalary,
         String status
 ) {
 }
